backend/requirements.txt
# Django & Django Ninja
Django==5.0.6
django-ninja==1.2.0
django-ninja-extra==0.21.0
django-ninja-jwt==5.3.0
django-cors-headers==4.3.1
django-celery-beat==2.6.0
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.9
django-redis==5.4.0

# Facebook API
facebook-sdk==3.1.0
requests==2.31.0

# AI & ML
openai==1.30.1
google-generativeai==0.8.3
google-genai>=1.0.0
anthropic==0.25.0
pillow==10.3.0

# Task Queue & Scheduling
celery==5.3.6
redis==5.0.3

# Storage & Media
boto3==1.34.84
django-storages==1.14.2
python-magic-bin==0.4.14  # Windows compatible

# Authentication & Security
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Environment & Configuration
python-decouple==3.8
python-dotenv==1.0.1

# Utilities
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
aiohttp==3.9.5
python-dateutil==2.9.0
pytz==2024.1

# Monitoring & Logging
sentry-sdk==2.0.1
loguru==0.7.2

# Testing (will be in requirements-dev.txt)